types: add default SyncBehavior
Summary:
# This diff

Implements a method for specifying no synchronization is needed, as this is the common case for clients issuing Eden thrift requests. You can read more about sync behavior in the eden.thrift in-code documentation

# Context

The EdenFS thrift service often suffers from Thrift "Queue Timeout" errors which indicate that the Eden daemon cannot keep up with the number of requests it's receiving and therefore does not dequeue Thrift requests from the queue fast enough. In the past, there hasn't been an easy/reliable way to trigger these Queue Timeout errors.

This diff stack adds a new command, `eden debug stress` that can be used to trigger Queue Timeout errors or simply stress test Eden to understand how it performs under certain workloads.

The command is meant to be easily extendable, so that we can easily plug in new Thrift endpoints to make them stress testable in the future. This diff adds the `getAttributesFromFilesV2` as the first endpoint to be stress testable. We may want to add `getFileInformation`, `globFiles`, etc in the future (we should definitely add `catFile`/`getFileContent` when that endpoint is added this half).

The subcommand takes the number of tasks, request amount, and mount point as optional arguments.

Reviewed By: jdelliot

Differential Revision: D71133462

fbshipit-source-id: 54f5e536e0d81c8ae3a65e95fc5966d1a4552f6f