Refactored connection code to allow for the daemon to become ready
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Following some of the buck2 pattern, we are enabling the ability to wait for a daemon to become ready by connecting on a spawned task. Presently, we just return ready.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Reviewed By: MichaelCuevas

Differential Revision: D71222328

fbshipit-source-id: 340a542317d0ebfcd27cc66af2d8ddbad5cb2da1