Reintroduce timeout for connecting to thrift (different from the actuall thrift client timeout for connection)
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

We used to have a timeout during connecting to the thrift server. This reintroduces that. It probably makes sense to apply this timeout to the entire execution of the the thrift method (connect + invoke), but that is not what we had.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Reviewed By: kavehahmadi60

Differential Revision: D71224063

fbshipit-source-id: 96a831746be6e7788773ab9935018511ec52d14d