Switched to new creation pattern for streaming thrift client
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

In earlier diffs, we needed to switch to the new creation pattern for the non-streaming client. Let's go ahead and do the same for the streaming client.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Reviewed By: MichaelCuevas

Differential Revision: D71220827

fbshipit-source-id: 7fd8c96c9886f57b45492231c8700acd563386e4