Impelment daemon running logic after connecting, but before invoking a thrift endpoint
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Added new logic to poll the daemon for up to 10s - validating it is running. Until we are caching and reusing connections this will add some trivial amount of time to all client lib (and therefore CLI) operations. Soon, we will add in reconnect/retry logic that will include reusing existing connections - which will be better than before.

# Next Steps:

* Continue to pull in more and more of the buck2 eden module - adapting it to fit into our specific needs

# Facebook:

Reviewed By: kavehahmadi60

Differential Revision: D71229436

fbshipit-source-id: be09e83b3d771f2f00b171e5e821108858474f4e