Update SelectMaxId query to return 0 when no rows
Summary:
The usages of this query (particularly in update preloaded graph subcommand) suggests that we expected this query to return an empty Vec if the table contains no rows which is why we have `unwrap_or(0)` in places to account for that situation.

However, XDB doesn't do that. Using `max` constraint over an empty table simply returns Null and deserializing the result into a `u64` errors out. This diff modifies that query so in cases where the table is empty, we return 0 as expected.

In 99% of cases, this should make no difference. But for a few anomalous repos with no entries in the commit graph edges table, this will fix the expected response.

Reviewed By: YousefSalama

Differential Revision: D71312724

fbshipit-source-id: 8a12d8506e518870bf2743cae7a913413d3041a8