tests: disable test-debugruntest.t on Windows
Summary: On Windows this test fails because `debugruntest` has problems launching itself, which causes the test to time out. It seems to be related to known problems of `debugruntest` that sometimes require adding `#inprocess-hg-incompatible` to tests. However, this specific test cannot have that setting enabled because otherwise it fails on Linux and macOS

Reviewed By: quark-zju

Differential Revision: D71324322

fbshipit-source-id: 3d6d5b29122f465a6150f7cd2c0f078683903494