Back out "config: make automerge support mobile config schema"
Summary:
D62892848 was landed as a productivity win for engineers using mobileconfig. Unfortunately, the merge driver algorithm assumes that the params inside of a config are alphabetized. That is not a true assumption and can lead to the mergedriver producing bad configs like:

D71053298

You can reproduce the error with: D71333447

The plan is to start updating the tools to alphabetize params and then re-enable when it is finished.

Reviewed By: zzl0

Differential Revision: D71335482

fbshipit-source-id: ea1eca2099f35702c9e7487500e6ad1f34c77818