Load the hg changeset earlier
Summary:
## This stack

Syncing fbsource fails because we try to upload trees for null changes.

## This diff

D71210512 changed the order of operation to derive and load the changeset in parallel with other fetches. That is good for performance, but in practice I couldn't really measure a performance difference.

More importantly though, this now gets in the way of further refactoring I'm doing, making it a premature optimization. Let's fix the fbsource bug first, then we'll benchmark again and optimize if needed.

Reviewed By: lmvasquezg

Differential Revision: D71392992

fbshipit-source-id: 6916d6d61e665bf34008e4fca7e1c9d7007b789c