queues: re-introduce bounded queues
Summary:
Add back bounded queues for the various stages of (non-eden) checkout. This includes the manifest diff, file/tree fetching, and writing the files out to disk. This avoids using unlimited memory when one phase (e.g. writing files to disk) is slow.

There is also a bounded queue in the http client, but that is controlled by a config since it is riskier in how it works. The part that is "too fast" is downloading file content, so currently we will still use unlimited memory during big checkout operations (until the http bounding is turned on).

This was previously reverted in D70366746 due to accidental perf regression. Crossbeam's unbounded channel pre-allocates the specified capacity, which were big (slow) allocations, even when fethcing only one tree/file. That issue is resolved by switching to a different channel implementation (flume).

I switched the channels from crossbeam to flume because flume:
  - supports sync and async (including Stream impl)
  - doesn't preallocate queue for bounded()
  - supports sync try_send()
  - no unsafe

Other options considered:
  - chadaustin's batch-channel (missing Stream impl and try_send, unsafe)
  - crossbeam-channel (pre-allocates bounded queue, no Stream impl)
  - kanal (pre-allocates bounded queue, unsafe)

This actually simplified the logic in the http receiver a lot. flume uses a single concrete type for bounded/unbounded/sync/async, so we don't need a wrapper type.

Reviewed By: quark-zju

Differential Revision: D71224226

fbshipit-source-id: 589ccb183a640809af0a914a573c8a1dcca199c8