make period inode unloading and periodic GC exclusive
Summary:
Two periodic tasks cannot unload and delete inodes at the same time.
On Windows, we don't have a periodic inode unload task, so we don't have any issue with the GC.
On Linux and macOS, we do have a periodic inode unload task, so we need to confirm that the periodic inode unloading is disabled when the GC is enabled.

Reviewed By: genevievehelsel

Differential Revision: D71352700

fbshipit-source-id: 2f5b0553a09bf5282b74467d5c318c3def0445ec