testing: update pythonXY.dll discover logic
Summary:
When running the locally built `sl`, its `pythonXY.dll` needs to
be in PATH. Otherwise, the `sl` program will silently exit, and
it's hard to debug.

Previously, the `pythonXY.dll` uses `sys.path` to discover its
path. However, `sys.path` no longer includes the directory of
`pythonXY.dll`. Let's also consider the `PATH`.

Since this is a hard-to-debug issue, I also made it an error so
this issue is easier to discover in the future.

Reviewed By: muirdm

Differential Revision: D71407367

fbshipit-source-id: 942688cbc58fd3689c11402de650e6dc4c16fadf