make tests compatible with git-lfs 3.6.0
Summary:
git-lfs checkout command used to discover LFS pointers by listing all the files in the repo and inspecting each one for being a pointer.

After [a change](https://github.com/git-lfs/git-lfs/commit/beae1146bda7e9fa712c3c73dacbcf9ed5067fd9) that came with 3.6.0 release the situation is different: git is relying on files tracked as LFS in .gitattributes file to narrow the scope. This was done to increase compatibility with sparse checkouts.

In our tests we didn't commit gitattributes files to the repo. We implicitly created it with `git lfs track` command but this was after the first checkout was already done.

The solution is to serve the valid gitattributes with lfs pointer bringing our test setup  much closer to setup to what actual repos using LFS do in production.

Reviewed By: RajivTS

Differential Revision: D71423752

fbshipit-source-id: a4dc9f5763151cf003afc4adb3a2432f5cd3f7db