use mononoke.server:prod rather than mononoke.server:stable in trinity tests
Summary:
We should use `prod` instead of `stable`.

Conveyor tags mononoke.server with `prod` when it actually pushes to prod (https://fburl.com/scuba/fbpkg_package_history/ktsnfyc5) which is what we really want.

`stable` is managed by some chronos job: https://www.internalfb.com/chronos/job/gp/99079200160620948, which lags with `stable` from actual prod release.

This increases the amount of time to land changes significantly but unnecessarily. E.g. when I have 2 diffs: one with `.thrift` changes and the other one where I use it, I need to wait for the former to get pushed and then tagger with `stable` just for the tests to pass. this can take days or week is we have a build broken for couple of days.

Reviewed By: andreacampi, sggutier

Differential Revision: D71475748

fbshipit-source-id: 24183ac27dcd39ca5289f97f6a86bc2dc8746772