subtree: add "import" subcommand for importing external repos
Summary:
## Context

A `subtree import` copies a tree of code from an external location (typically Git repos) into the monorepo.
Internally this appears very similar to a large add of a tree of files, but it will contain additional metadata
about the source of that code, i.e. the repo name, commit hash and path that the tree was copied from.

## This diff

Adds the initial version of `subtree import`, which clones the external repo into a temporary directory and
then moves the specified from_path to its corresponding to_path.

Reviewed By: quark-zju

Differential Revision: D71245619

fbshipit-source-id: 9b0d0ea3256bf0c582737515ba1d99e8b2d23330