add integration test
Summary:
Adding integration test showing that the bundle-uri command on gitprotocol works.

I'm also pulling the meta vs oss functionality behind the bundle-uri crate itself so we don't have to do any weirdness with build-time macros outside of the crate. This also allows us to have the LocalFSBundleGenerator and the CDN one have different signatures for the `new` function (I'm actually removeing it altogether for the LocalFS-one as it's not needed).

Reviewed By: RajivTS

Differential Revision: D70923438

fbshipit-source-id: 7870aef379768333ad1ebbe86f31e90fb1315d9a