ODS cleanup: rename changeset series
Summary:
## This stack

Update or add logging and stats that we need to troubleshoot the performance of modern sync.

## This diff

We have a typo in the name of a time series (`procesed` vs `processed`). Changing that will lose older data of course, so we might as well use the opportunity to make more changes:

* use a unique prefix instead of clashing with `manager.rs`;
* prefer dots to underscores for key prefixes that identify the same entity;
* renamed variables to match.

Reviewed By: mzr

Differential Revision: D71474867

fbshipit-source-id: 24f692c26e6cb6bc78255051d48c26cdd1c2d3d4