mercurial/types: parse large manifests on a blocking thread
Summary:
Large manifests take some time to parse, and block the executor if this is done on an executor thread.  Small manifests, however, are fine, and we don't want to incur the overhead of shuffling the parsing over to another thread in that case.

Compromise by picking an arbitrary manifest size (1MB) and parse larger manifests on a blocking thread.

Reviewed By: zzl0

Differential Revision: D71473776

fbshipit-source-id: 631e8efb1438878fe0874de872ec42d3dbd3d84f