tools/admin: make rederivation rederive all selected derived data types
Summary:
Use of `HashSet<ChangesetId>` for the implementation of `Rederivation` in `monad derived-data derive --rederive` means we can only ever rederive one type, as that type will remove the changeset id from the set, preventing other types from rederiving for that changeset.

Switch instead to using a per-derived-data-type set, so we can rederive multiple types for the same changeset.

Reviewed By: liubov-dmitrieva

Differential Revision: D71473778

fbshipit-source-id: c9a070f6253c0d11f7805aeed01be9fe25c4d09d