Open the repo unredacted
Summary:
## This stack

Update or add logging and stats that we need to troubleshoot the performance of modern sync.

## This diff

This will bypass redaction and just sync the actual content.

After discussing it in the pod sync, this is the more useful approach because it ensures the AWS repo can be used (in the future) for backup & restore.

Reviewed By: clara-9

Differential Revision: D71547665

fbshipit-source-id: db081a2de5fdfadc57d007eea56608ddae7fde68