setting up CommitInforField to take in customized field
Summary:
**Project Proposal:** https://docs.google.com/document/d/1zMkh3jnXHVzBQn3i7pzdQR4u5MwADh8xuYd9es7juvA/edit?tab=t.0#heading=h.i3ds4cczfhw4

**Design:**
https://www.figma.com/design/szRjvQIQBiHcizuxY5e5TW/Diff-Rollback-Field?node-id=19-15073&p=f&t=HQmkhPzBZnbyETDI-0

**Context:**

I realized that the CommitInfoField and all of the CommitMessageFieldsSchema are either a field or textarea. This makes sense as all of these are representative of data attached the `commit message` of a commit. However, I would want to create a custom component to show something similar to the design on Phabricator end (see design doc).

Is what I'm doing here ok since ISL is shared with OSS? I will replace the TODO placeholders with Internal Components. Or is there something I should do in the CommitInfoField level to hide these internal logic? Or some other completely different way that I should handle this? Open up for some thoughts. cc. evangrayk

**Not going to land this Diff as is. Just putting it up to ask for some feedbacks.**

Reviewed By: evangrayk

Differential Revision: D71455480

fbshipit-source-id: ca0e05ae08f186dce5844fb4067bf39f67b74a30