In benchmark mode, use in-memory MutableCounters
Summary:
## This stack

Update or add logging and stats that we need to troubleshoot the performance of modern sync.

## This diff

This lets us avoid affecting production; and will also allow us to introspect how far along we got in a benchmark run.

With this diff it is now safe to run a benchmark; the only interaction with the target repo are lookups.

Reviewed By: lmvasquezg

Differential Revision: D71563858

fbshipit-source-id: 8006612ef99ee59113fd0f68c9ebb482cce9c257