Move rewrite_commit to separate module
Summary:
## This stack
Starting the significant refactoring to improve the cross-repo codebase and its abstractions.
There are so many things to do, that's it's honestly hard to be brief...

I'll try to summarize the current state and the end state.

## Current state
We have (1) `commit_transformation` and (2) `cross_repo_sync` crates. The latter has functions that were copied from the former, along with logic to handle git_submodules (e.g. expand, compact, etc).

`cross_repo_sync` is very bloated and unorganized, e.g. there are circular dependencies between modules and many modules contain code from multiple abstraction levels.

## End state
We'll have the same two crates, but they will be put in the mononoke features directory and will have specific responsibilities, which a lot of work will be required to enforce in practice.

- commit_transformation:
  - Will handle the bonsai-level work of rewriting commits, rewriting paths.
  - Will have the git submodule expansion logic (currently in cross-repo-sync), since it's applied at the bonsai level.
  - Will expose `rewrite_commit` and auxiliar types.
  - This is lower level and will **not take a target repo** as input. Source repo is needed for reading.
- cross_repo_sync
  - Will handle repo-level cross-repo logic, e.g. commit sync configs, small/large repo stuff, sync normal or via pushrebase.
  - Will expose functions like `sync_commit`, `unsafe_sync_commit_via_pushrebase`,

## This diff

Starts the work of re-structuring everything by moving `rewrite_commit` to a separate module.
These are "lower-level" functionalities used by the rest of cross-repo-sync (e.g. commit_syncer) and it will be extracted to its own crate together with git_submodules logic, to then be moved to commit_transformation.

rewrite_commit was copied from `commit_transformation/lib.rs` (see same doc comments) in D28259214 and, in the end, we'll remove this copy and consolidate it with the one in commit_transformation crate.

NOTE: **I'm trying to do things incrementally, so most of this will be removed to a separate crate later on, to then be merged with the commit_transformation crate. But doing this requires a lot of small things to be done in between, so I decided to do it in multiple diffs that are easier to review.**

Reviewed By: mitrandir77

Differential Revision: D70781050

fbshipit-source-id: 399c39ea4a3f8c6e1f40c1789856348801ed4bad