Split reporting module into separate rust crate
Summary:
# This stack
## **Please see D70781050 before reviewing this to get context on the plan and the goal.**

## This diff
As mentioned in previous diffs, I'll be extracting commit_rewrite into a separate crate in the next few diffs. But it depends on the reporting code, so let's extract this to a separate create as well to unblock the refactoring.

This crate should probably be deleted, but doing that now will take some time and I'm prioritizing improvements at a high level right now, so I added a TODO to get back to this later.

Reviewed By: mitrandir77

Differential Revision: D70790442

fbshipit-source-id: 2cf779d5f89eb0b9aae8876dfe7c8fe874f28a81