Define Repo trait for commit_rewrite modules
Summary:
# This stack
## **Please see D70781050 before reviewing this to get context on the plan and the goal.**

## This diff

The git submodules logic and rewrite_commit are lower level components that work at the bonsai level, so they will be moved to a separate crate.
In order to do that, we should stop using the same Repo facet definition that's used **across the entire cross-repo-sync** codebase.
This diff will define a new trait, with the strictly necessary repo attributes for the rewrite_commit operation.

Reviewed By: mitrandir77

Differential Revision: D70783247

fbshipit-source-id: ada9e80a767702e1d19a44c744de7f5aedf6422b