Refactored EdenFsClient into two structs/impls - one non-streaming, one streaming
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Refactored EdenFsClient into two structs/impls - one non-streaming, one streaming and placed these into the client mod. Overall, this is preferred, but there is one instance where a associated function needs both types. This is probably the wrong pattern and we can address in subsequent diffs.

Tried to create the appropriate trait, but ran into some trouble with the Thrift types not implmenting Sync. This should be solvable, but after several hours I put it on hold for now.

We can follow this same pattern to break apart EdenFsConnector as well. On negative of this, is that each client can have the max number of connections - we should be able to mitigate that by sharing the semaphore for the default and/or wrapping these types up in client/connection manager which house the semaphore.

# Next Steps:

* Add in throttling via a semaphore to keep from overwhelming EdenFS.
* Refactor duplicative code in connection and client
* Introduce a singleton for the client - as there should only ever be one

# Facebook:

Reviewed By: MichaelCuevas

Differential Revision: D71433068

fbshipit-source-id: 17c39dbcd42ba8e426f77a4d9cc96718d10647cb