Removed logic to wait for daemon on connect
Summary:
# Context:

We are working to refactor the EdenFs CLI client library into a supported XFN consumable library to avoid code duplication and sub-optimal usage patterns.

# This Diff:

Currently, we wait for the daemon to startup on every connection. This is overkill as 99%+ of the time the daemon will be running and ready. We will advise clients of this library to test any daemon readiness (e.g. get_daemon_info) where they required. This diff removes that test from our connection flow.

# Next Steps:

* Introduce a singleton for the client - as there should only ever be one

# Facebook:

Reviewed By: genevievehelsel

Differential Revision: D71685074

fbshipit-source-id: 93997a11b2cecd65c48973c88feef7dc61729f45