histedit: fix bad dirstate after "fold" commit
Summary: After folding a commit, we weren't update the dirstate properly. We were using `ctx.markcommitted(n)`, but that only considers the delta between ctx and "." (i.e. it assumes a normal "commit" operation). For "fold", the new `ctx` is replacing ".", so the dirstate manipulations were not correct. Instead, perform more of a "reset" operation, marking things in the dirstate as either clean or deleted, depending on whether they exist in the folded together ctx.

Reviewed By: zzl0

Differential Revision: D71432665

fbshipit-source-id: 0cdb50578b5195b64ae8e5f805bd0436aeaf36dc