Integration test highlighting failure in shallow clone with submodule
Summary:
When shallow cloning a repo from Mononoke Git, we fetch all the objects from the root trees of the boundary commits. Before deciding to include them in the packfile, we filter out the ones specified by the filter criteria. One of the filter criteria is the size.

For most entries in the Git Tree, the size constraint makes sense. As long as the entry is a blob or a tree, we look at the size to decide if it should be included or not. However, when the entry is a submodule, the concept of size makes sense. The current version of the code doesn't make an explicit bypass for submodule entry but instead tries to fetch its size and fails. This integration test highlights that problem

Differential Revision: D71734349

fbshipit-source-id: 6e310cfba3096e88a8afad6785f9904d3dec5329