Fix bug in implementation of shallow in Mononoke Git server
Summary: As demonstrated in the previous diff, shallow clone of a submoduled repo can fail depending on the shallow boundary. This diff adds an explicit check before applying the size filter for the shallow boundary. If the object is in object is a submodule, do not perform the size check.

Differential Revision: D71735051

fbshipit-source-id: 5dff655e87b05b82fb3b3909ee47bdf991922067