Assert that the bonsai id matches after upload
Summary:
I have been [troubleshooting](https://fb.workplace.com/groups/1708850869939124/permalink/1929183197905889/) an issue where we didn't copy everything–in effect, the changesets were not identical after all.

This was hard to debug after the fact; it would be better if we failed right away. Luckily that is really cheap and easy.

Reviewed By: markbt

Differential Revision: D71738546

fbshipit-source-id: 9b23024e539f348b3b7e73d7cedf5af9090d97b2