Update fbpkg config to exclude dwp files
Summary:
## This Diff
Added an argument to build edenfs binaries without the debug info and get rid of the dwp files.
This decreases the package size significantly, thus speed up the installation process.

ref: https://fburl.com/wiki/bs2r1ji6

## Context
In this diff (D71420020), J.T. noticed that we are shipping dwp files together with the eden binaries in fbpkg, which caused the fbpkg size to be quite large. This was a concern since we were adding EdenFS into the list of fbpkgs a sandcastle worker would fetch when initializing. Although the content of fbpkg is cached, the first fetch is always going to be affected whenever the version of the package is bumped.

Reviewed By: MichaelCuevas

Differential Revision: D71572724

fbshipit-source-id: 8450308bc1f21bd231ba9e39b177b97058af2d25