do not use draftonly identity for git server
Summary:
The thing that really prevents the clients from pushing is the DraftOnly identity (see the test plan of D63692445).
This shouldn't be set by modifying the trusted/untrusted variable in metadata. Client either is in the untrusted network or not. How we use the is_untrusted bit is another story, and that's what I'm doing here - I'm moving the decision to the place where we use this.

This is driven by the next diff, where we want to still allow pushes in vpnless (aka untruested network), but make other decisions based on is_untrusted bit.

Reviewed By: RajivTS

Differential Revision: D71698913

fbshipit-source-id: 0eebe85538883986205840c3a86ea9f7a75ad93d