update client_untrusted in metadata everywhere except git server
Summary:
I believe the current implementation (to allow the git clients to push from vpnless to mononoke-git-server) is unsafe. The clientinfo is controlled by the client, and therefore it's unsafe - they can set it to whatever they want. This way anyone can make their client "safe" by setting the clientinfo's entrypoint to mononoke-git-server manually.

Let's have this set on the backend, per-service.

This diff is for demonstration only. I'm changing the solution to another one in subsequent diff.

Reviewed By: RajivTS

Differential Revision: D71697463

fbshipit-source-id: 88cc8b883d73d7de26faf434dc681ce34c3bac18