hooks: add hook for limiting subtree copy source sizes
Summary:
Add a new hook that lets us limit the size of subtree copy sources, and the depth of the destination.

There are two aspects of "size" we care about when it comes to limiting subtree operations: the size of the source, and the size of the covered portion of the repository at the destination.  For source size we limit the total number of files located at the source.  For the destination, we require that it has a minimum path depth.

Reviewed By: andreacampi

Differential Revision: D71842445

fbshipit-source-id: c89713185eb56975b48e9af91a6176817f870274