Add Goto to right-click menu
Summary:
See https://fb.workplace.com/groups/isl2.0/posts/1612234842746926

Basically the Goto button is on the far right of the ISL. This makes it difficult to click if ISL is narrow. Now, you can goto by right clicking.

Alternatively, double clicking on the commit may be a good way to goto a commit too.

Reviewed By: evangrayk

Differential Revision: D71894646

fbshipit-source-id: 2472bc7f96e0530ba2d11c7ffc82cc43dd704ac7