Merge commit_rewrite into commit_transformation crate
Summary:
# This stack
## **Please see D70781050 before reviewing this to get context on the plan and the goal.**

## Context for this diff
The `rewrite_commit` from cross_repo_sync (now in `commit_rewrite` crate) was copied from commit_transformation, but it shouldn't have been.

In the next few diffs I'll merge commit_rewrite crate into commit_transformation.
After some clean up, this crate will then be moved to the mononoke/features directory.

## This diff
Finally merges the newly extracted `commit_rewrite` crate into the `commit_transformation` crate.

Most of the diff just moves code around, but there were some actual changes that had to be made **in this diff** to consolidate the two `rewrite_commit` functions into a single one.

I deleted a struct called `MoverMultiMover` in the `rewrite.rs` file. This was just a wrapper type that was created to implement `MultiMover` for the type `Mover`. I removed the need for it by implementing `MultiMover` for all `Arc<dyn Mover>`. I tried to do it in a separate diff before, but I hit orphan errors because I can't implement traits where both the trait and the type for which I implement are from outside the crate.
Since in this diff everything is moved to `commit_transformation`, I added the implementation to the `types.rs` module.

Differential Revision: D71473716

fbshipit-source-id: afee1d2da35082826faba5da8450514c4882294c