scmutil: simplify addremove arg handling
Summary:
Simplify to only accept "opts" instead of opts and/or dry_run and/or similarity. The particularly confusing/wrong part was "similarity" in opts was a different format to the `similarity` arg (the former was [0, 100], the latter was [0, 1.0]).

As much as I hate the nebulous "opts" argument, it's better than having opts and other args that overlap, sometimes.

I also got rid of the unused "prefix" argument.

Reviewed By: sggutier

Differential Revision: D71487875

fbshipit-source-id: 8b42283486c1e7dc267c34ff71b76a516f288f83