support exporting tar file to stdout
Summary:
Now you can specify `--output -` to write the tar contents to stdout instead of a filesystem path.

Exporting to stdout is useful when dealing with large exports you don't want to (or don't have room to) write to disk. It is also nicely composable with `| gzip` or various other unix pipelines.

Reviewed By: andreacampi

Differential Revision: D71936432

fbshipit-source-id: 027fed50931cd4d1c3cc23958354161ffb7ae155