render output to stderr if stdout isn't tty
Summary:
Add a fallback so we write tty output to stderr if stdout is not a tty (otherwise the output is dropped).

This allows, for example, progress output to still show up in the terminal if the scsc command is writing its "real" output to stdout (used in next diff).

Reviewed By: andreacampi

Differential Revision: D71936431

fbshipit-source-id: 81bfbe80d33503a91632fa21277214bebf25c14c