tests: flag failing Windows test
Summary:
There recently were some changes to how symlinks work in D71445977. That diff broke the `clone` command for sparse profiles on Windows. Although this is a very niche use case, the test was originally added as a test of sorts to make sure other Python commands also worked on Windows with sparse profiles with symlinks.

Instead of deleting this test, this diff just flags it in case the use case warrants an investigation later.

Reviewed By: muirdm

Differential Revision: D71953331

fbshipit-source-id: 0c0021b034d2154259c9092aafa330cc765ef961