FsChannel: pipe longRunningFSRequestThreshold through
Summary:
# This diff

Pipes the new config value through to all the FSChannel implementations. This will allow us to configure the LongRunningFSRequest threshold without shipping a new binary.

NOTE: the reason I didn't pipe the EdenConfig through directly is because we don't want to reread the config on every single NFS/FUSE/PrjFS request that we handle. Instead, we'll read the config value once at mount time (when we create the FSChannels), and pass in the raw value to the RequestContext constructor.

# Context

On macOS, we're seeing frequent NFS timeout errors. We have theories about why the timeout errors are occurring, but we have no concrete evidence for these theories. This diff stack is an attempt to rule out (or confirm) one theory: NFS timeouts are currently "out of our control" because the majority of the timeout period is spent before Eden even accepts the NFS request and starts working on it.

We can test this by recording the entire duration of the NFS request (from when we dispatch the request all the way until we respond). If we confirm that some requests are running for a long period of time after we dispatch, then we'll confirm that the problem lies within Eden itself. If we see that no requests are running for a long duration within Eden, then the time must be spent before Eden dispatches the request (in a queue), or there's a bug in the NFS client where some requests are never delivered to EdenFS.

Reviewed By: jdelliot

Differential Revision: D71908866

fbshipit-source-id: af386873875e30d107aa80f6a916d401177bba68