RequestContext: use threshold specified by config
Summary:
# This diff

Sets the LongRunningFSRequest threshold via the EdenConfig value. This will allow us to set different thresholds via dynamic config instead of being forced to ship a new binary.

# Context

On macOS, we're seeing frequent NFS timeout errors. We have theories about why the timeout errors are occurring, but we have no concrete evidence for these theories. This diff stack is an attempt to rule out (or confirm) one theory: NFS timeouts are currently "out of our control" because the majority of the timeout period is spent before Eden even accepts the NFS request and starts working on it.

We can test this by recording the entire duration of the NFS request (from when we dispatch the request all the way until we respond). If we confirm that some requests are running for a long period of time after we dispatch, then we'll confirm that the problem lies within Eden itself. If we see that no requests are running for a long duration within Eden, then the time must be spent before Eden dispatches the request (in a queue), or there's a bug in the NFS client where some requests are never delivered to EdenFS.

Reviewed By: jdelliot

Differential Revision: D71908865

fbshipit-source-id: ca0d8580d75c073c520e9e9b72ea24fc29c271f0