Avoid busy polling on the BUL when caught up
Summary:
## This stack

Small tweaks for the admin repo.

## This repo

While try to fix the ODS counter for admin I realized that when we have no entries to process (i.e. when caught up) we fail to sleep at all. This uses a lot of CPU, but probably also generates a fair number of DB queries.

Reviewed By: lmvasquezg

Differential Revision: D71975890

fbshipit-source-id: 6a3d0ca07b29ead2a8ddf28fd51c9c5a5873527b