subtree: deprecate subtree.copy-max-file-count config
Summary:
This diff makes "subtree copy" use the "subtree.max-file-count" config
same as other commands such as "subtree import" and "subtree merge". The
default limit has been set to 100,000, which should be sufficient for
most cases.

Sapling also plans not to expose this config to users, as it can generate
arbitrarily large copies.

Reviewed By: muirdm

Differential Revision: D71736782

fbshipit-source-id: f9ccce8e25463af403a7c5608f3bc31661d8ae93