stackEdit/diffSplit: add a way to export a diff of a commit as json
Summary:
This is intended to send a "diff" to a service to calculate a split
(of the diff), by seleting a subset of the changed lines.

We calculate the diff client-side so the server sees the exact diff as a human
sees, and can only choose the split based on the same diff. This is subtle.
Example:

Diff 1:

   a
  +b
  +c
   b

Diff 2 (same left/right side content, but different diff):

   a
   b
  +c
  +b

Valid splits of Diff 1 and Diff 2 are different. For example:

   a
  +b
   b

is a valid split in Diff 1, but not in Diff 2.

Reviewed By: evangrayk

Differential Revision: D71830003

fbshipit-source-id: 7ea0776724c6120de63ffdcbc038f2cd44e9fbdd