progress: add "noop" mode to disable progress work
Summary: Registering progress bars is relatively expensive (RwLock), so we want to avoid it when progress is not used. It isn't convenient to add "if enable_progress" to all users of progress bars, so let's add a no-op mode to ProgressBar/Registry that short circuits the work of creating and udpating progress bars (using an AtomicBool).

Reviewed By: sggutier, zzl0

Differential Revision: D72003224

fbshipit-source-id: 627276177e9338be00e1b1465758aec7f6c50ab8