Swap entity & key in query
Summary:
This diff addresses a bug where we saw the counters come back empty as we passsed the arguments to the function in the wrong order, there are several ways to address this. We likely want to be passing around a struct that contains the different fields instead of strings.

Because the OdsCouters ar eused both for ratelimiting and in the Jupiter for external dependencies deployment I need to have a think about the right naming, and thread it thoughout the code. Let's land the fix for now.

Reviewed By: YousefSalama

Differential Revision: D72182790

fbshipit-source-id: b8a47c82d6603a16a14482e685aabacb08399bfb