cmdlib/sharding: don't immediately fail after cancelling setup
Summary:
When we call `.abort()` on the setup `JoinHandle`, it's extremely likely that polling the join handle will result in a tokio `JoinError` for that join handle because the task was cancelled.

There's no point in surfacing this as an error that stops the whole process.  It was cancelled successfully.

Reviewed By: RajivTS

Differential Revision: D72176112

fbshipit-source-id: 463884a75070cb9305624697e20aab113440df13