utils: add infallible version of path_from_bytes
Summary:
# This diff

Adds a utility function for converting bytes (that may contain non-UTF-8) into a lossy path.

# Context

In some cases, we can guarantee that the bytes we're working with are valid UTF-8 (such as when the Eden daemon passes us path results). Therefore, we can confidently convert to a lossy result without any fear of the conversion actually being lossy.

In other cases, we actually don't care about the conversion being lossy. This function can also be used in that case.

Reviewed By: jdelliot

Differential Revision: D72091595

fbshipit-source-id: 2dae6bc24f3293c34d77c4dd3c018b31d6cfacd7