types: use new TryIntoFileAttributesBitmask trait
Summary:
# This diff

Uses the new trait that was added in the previous diff. The new trait allows us to avoid having duplicate functions for different forms of FileAttributes

# Context

The EdenFS clients that we have the most trouble with are those that do some flavor of recursively enumerating directories, asking for aux data about each directory/file, and then optionally reading file content for some of the directories that were enumerated[1]. These workflows are a great stress test for EdenFS, because thousands (and potentially millions) of files/directories are being queried in a short amount of time.

Before this diff stack, we didn't have a good way of simulating these workflows without running the aforementioned tools directly. These tools often have other caching implemented in them, so it's also hard to get a consistent repro of performance issues[2].

This stack introduces a recursive readdir stress subcommand so that we can test how Eden performs when recursively enumerating directories and fetching aux data (tree and file aux data) for arbitrary sections of the repository. The stress subcommand also allows you to specify how much concurrency to use when performing the recursive readdir so that we can better understand how EdenFS performs when performing many fetches at once.

Reviewed By: jdelliot

Differential Revision: D72091594

fbshipit-source-id: d5bdaf38959cbe1a87a2dc558a197a6c840139e9