add FileAttributes wrapper
Summary:
# This diff

Adds FileAttributes structs and trait impls and replaces the usage of thrift_types::edenfs::FileAttributes with the new struct

# Context

The EdenFS clients that we have the most trouble with are those that do some flavor of recursively enumerating directories, asking for aux data about each directory/file, and then optionally reading file content for some of the directories that were enumerated[1]. These workflows are a great stress test for EdenFS, because thousands (and potentially millions) of files/directories are being queried in a short amount of time.

Before this diff stack, we didn't have a good way of simulating these workflows without running the aforementioned tools directly. These tools often have other caching implemented in them, so it's also hard to get a consistent repro of performance issues[2].

This stack introduces a recursive readdir stress subcommand so that we can test how Eden performs when recursively enumerating directories and fetching aux data (tree and file aux data) for arbitrary sections of the repository. The stress subcommand also allows you to specify how much concurrency to use when performing the recursive readdir so that we can better understand how EdenFS performs when performing many fetches at once.

Reviewed By: jdelliot

Differential Revision: D72091599

fbshipit-source-id: 8053cee717ec128035235ef780a4d50fb52ced8a