Rename auxDataCache to blobAuxDataCache
Summary:
Rename auxDataCache to blobAuxDataCache so it's more clear that which in-memory cache should be used to serve BlobAuxData or TreeAuxData.

## Context
Currently we fetch TreeAuxData directly from the backing store without caching. We want to add cache support. The cache will be consist of the in-memory cache (in ObjectStore) as well as the local store.

Reviewed By: jdelliot

Differential Revision: D72259841

fbshipit-source-id: 13b715f6e65c83b9c2c8cdb5119acb8ccffb8660